added teams file

+++ questioneer/Core/Entities/Team.cs
@@ -3,6 +3,6 @@
     public class Team
     {
         public string Name { get; set; }
-        public int Channel { get; set; }
+        public long Channel { get; set; }
     }
 }
